feat(notion): Add Notion integration schemas and update dependencies
- Introduced new schemas for Notion integration, including block and page schemas using Zod for validation.
- Added @notionhq/client dependency to package.json.
- Updated tools and types to reference new schema definitions.
- Removed common.ts as its contents have been integrated into schema.ts for better organization.
