+++ agent-packages/packages/notion/package.json
@@ -22,6 +22,7 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
+    "@notionhq/client": "^3.0.1",
     "axios": "^1.8.2"
   }
 }
