+++ agent-packages/packages/notion/src/types/args.ts
@@ -16,7 +16,7 @@ import {
   searchSchema,
   updateBlockSchema,
   updatePagePropertiesSchema
-} from './schema';
+} from '../schema';
 
 export type SearchArgs = z.infer<typeof searchSchema>;
 export type AppendBlockChildrenArgs = z.infer<typeof appendBlockChildrenSchema>;
