+++ agent-packages/packages/notion/src/types/index.ts
@@ -6,5 +6,3 @@ export interface NotionConfig extends BaseConfig {
 
 export * from './args';
 export * from './responses';
-export * from './schema';
-export * from './common';
